user_logic_controler

+++ models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	UserID int16 `db:"user_id"`
+	UserID int64 `db:"user_id"`
 	Username string `db:"username"`
 	Password string `db:"password"`
 	Token string
